refactor: implement config hot reload functionality
