+++ master/master.go
@@ -16,10 +16,13 @@ package master
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"math"
 	"math/rand"
 	"net"
+	"os"
+	"path/filepath"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -27,6 +30,7 @@ import (
 
 	"github.com/coreos/go-oidc/v3/oidc"
 	"github.com/emicklei/go-restful/v3"
+	"github.com/fsnotify/fsnotify"
 	"github.com/gorse-io/gorse/common/log"
 	"github.com/gorse-io/gorse/common/monitor"
 	"github.com/gorse-io/gorse/common/parallel"
@@ -52,6 +56,8 @@ import (
 	"golang.org/x/oauth2"
 	"google.golang.org/grpc"
 )
+
+const configReloadDebounce = time.Second
 
 type Datasets struct {
 	rankingDataset  *dataset.Dataset
@@ -149,6 +155,116 @@ func NewMaster(cfg *config.Config, cacheFolder string, standalone bool, configPa
 		cancel:    func() {},
 	}
 	return m
+}
+
+func (m *Master) applyRecommendOverride(cfg *config.Config) error {
+	metaStr, err := m.metaStore.Get(meta.RECOMMEND_CONFIG)
+	if err != nil && !errors.Is(err, errors.NotFound) {
+		return errors.Trace(err)
+	}
+	if metaStr == nil {
+		return nil
+	}
+	if err = json.Unmarshal([]byte(*metaStr), &cfg.Recommend); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+func (m *Master) reloadConfigFromFile() error {
+	newConfig, err := config.LoadConfig(m.configPath)
+	if err != nil {
+		return errors.Trace(err)
+	}
+	if err = m.applyRecommendOverride(newConfig); err != nil {
+		return errors.Trace(err)
+	}
+	if err = newConfig.Validate(); err != nil {
+		return errors.Trace(err)
+	}
+
+	m.ConfigMutex.Lock()
+	m.Config = newConfig
+	m.RestServer.Config = newConfig
+	m.ConfigMutex.Unlock()
+
+	select {
+	case m.scheduled <- struct{}{}:
+	default:
+	}
+	return nil
+}
+
+func (m *Master) watchConfigFile(ctx context.Context) {
+	if m.configPath == "" {
+		return
+	}
+	absPath, err := filepath.Abs(m.configPath)
+	if err != nil {
+		log.Logger().Error("failed to resolve config path", zap.String("path", m.configPath), zap.Error(err))
+		return
+	}
+	if _, err = os.Stat(absPath); err != nil {
+		log.Logger().Warn("skip watching config file", zap.String("path", absPath), zap.Error(err))
+		return
+	}
+	watcher, err := fsnotify.NewWatcher()
+	if err != nil {
+		log.Logger().Error("failed to create config watcher", zap.Error(err))
+		return
+	}
+	defer watcher.Close()
+	if err = watcher.Add(filepath.Dir(absPath)); err != nil {
+		log.Logger().Error("failed to watch config directory", zap.String("path", absPath), zap.Error(err))
+		return
+	}
+	log.Logger().Info("watch config file", zap.String("path", absPath))
+
+	var timer *time.Timer
+	var timerC <-chan time.Time
+	for {
+		select {
+		case <-ctx.Done():
+			if timer != nil {
+				timer.Stop()
+			}
+			return
+		case event, ok := <-watcher.Events:
+			if !ok {
+				return
+			}
+			eventPath, err := filepath.Abs(event.Name)
+			if err != nil || eventPath != absPath {
+				continue
+			}
+			if event.Op&(fsnotify.Write|fsnotify.Create|fsnotify.Rename|fsnotify.Chmod) != 0 {
+				if timer == nil {
+					timer = time.NewTimer(configReloadDebounce)
+				} else {
+					if !timer.Stop() {
+						select {
+						case <-timer.C:
+						default:
+						}
+					}
+					timer.Reset(configReloadDebounce)
+				}
+				timerC = timer.C
+			}
+		case err, ok := <-watcher.Errors:
+			if !ok {
+				return
+			}
+			log.Logger().Error("failed to watch config file", zap.Error(err))
+		case <-timerC:
+			timerC = nil
+			if err = m.reloadConfigFromFile(); err != nil {
+				log.Logger().Error("failed to reload config file", zap.String("path", absPath), zap.Error(err))
+			} else {
+				log.Logger().Info("reloaded config file", zap.String("path", absPath))
+			}
+		}
+	}
 }
 
 // Serve starts the master node.
